Substituição de dropdowns especificos de itens e métodos de pagamentos para dropdown genérico;
